Simply how the client kwargs are populated
The changes introduced in 96dbe16 removed the necessity for tracking the
host outside of the rest of the arguments passed to the client. The
simplified logic drew attention to how `kwargs` was being populated.
This moves the host into `kwargs` and populates most other values as
part of the literal assignment.